frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
paramiko = "^2.7.2"
psycopg2 = "^2.8.6"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.2"
autopep8 = "^1.5.5"
pytest = "^6.2.2"
pytest-html = "^3.1.1"
pytest-cov = "^2.11.1"
Sphinx = "^3.5.3"
sphinx-rtd-theme = "^0.5.1"
coverage = {extras = ["toml"], version = "^5.4"}

# Pytest
[tool.pytest.ini_options]
addopts = "--doctest-modules --html=tests_output/report.html --self-contained-html --cov --cov-report=term --cov-report=html"
cache_dir = "tests_output/.pytest_cache"

# Testing coverage
[tool.coverage.run]
data_file = "tests_output/.coverage"
source_pkgs = ["frontend"]
branch = true
[tool.coverage.report]
skip_empty = true
fail_under = 70
[tool.coverage.html]
directory = "tests_output/coverage_report"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"